Refactor all notifiers in a single notifier
