+++ apps/transport/lib/jobs/promote_reuser_space_job.ex
@@ -11,25 +11,9 @@ defmodule Transport.Jobs.PromoteReuserSpaceJob do
 
     {:ok, _} =
       contact.email
-      |> Transport.PromoteReuserSpaceNotifier.promote_reuser_space()
+      |> Transport.UserNotifier.promote_reuser_space()
       |> Transport.Mailer.deliver()
 
     :ok
-  end
-end
-
-defmodule Transport.PromoteReuserSpaceNotifier do
-  @moduledoc """
-  Module in charge of building the email.
-  """
-  use Phoenix.Swoosh, view: TransportWeb.EmailView
-
-  def promote_reuser_space(email) do
-    new()
-    |> from({"transport.data.gouv.fr", Application.fetch_env!(:transport, :contact_email)})
-    |> to(email)
-    |> reply_to(Application.fetch_env!(:transport, :contact_email))
-    |> subject("Gestion de vos favoris dans votre espace réutilisateur")
-    |> render_body("promote_reuser_space.html")
   end
 end
